refactor: extract JointImpWrapper from JointHierarchyManager (650→464 lines)
Extract the JointImpWrapper inner class (~180 lines) into a package-private
top-level class. Pass a BooleanSupplier for the isSims() check instead of
closing over the enclosing class. Public API unchanged.

Closes #626
